+++ enginetest/uuidtest/doc.go
@@ -1,0 +1,2 @@
+// Package uuidtest provides utilities for working with UUIDs in tests.
+package uuidtest